+++ chat-client/src/contracts/serverContracts.ts
@@ -29,6 +29,10 @@ import {
     ListConversationsParams,
     CONVERSATION_CLICK_REQUEST_METHOD,
     ConversationClickParams,
+    McpServerClickParams,
+    ListMcpServersParams,
+    LIST_MCP_SERVERS_REQUEST_METHOD,
+    MCP_SERVER_CLICK_REQUEST_METHOD,
     GET_SERIALIZED_CHAT_REQUEST_METHOD,
     TAB_BAR_ACTION_REQUEST_METHOD,
     TabBarActionParams,
@@ -57,6 +61,8 @@ export type ServerMessageCommand =
     | typeof FILE_CLICK_NOTIFICATION_METHOD
     | typeof LIST_CONVERSATIONS_REQUEST_METHOD
     | typeof CONVERSATION_CLICK_REQUEST_METHOD
+    | typeof LIST_MCP_SERVERS_REQUEST_METHOD
+    | typeof MCP_SERVER_CLICK_REQUEST_METHOD
     | typeof TAB_BAR_ACTION_REQUEST_METHOD
     | typeof GET_SERIALIZED_CHAT_REQUEST_METHOD
     | typeof PROMPT_INPUT_OPTION_CHANGE_METHOD
@@ -89,5 +95,7 @@ export type ServerMessageParams =
     | FileClickParams
     | ListConversationsParams
     | ConversationClickParams
+    | ListMcpServersParams
+    | McpServerClickParams
     | TabBarActionParams
     | GetSerializedChatResult